+++ packages/ocean-react-formik/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.39.0](https://github.com/ocean-ds/ocean-web/compare/v1.38.3...v1.39.0) (2022-12-13)
+
+**Note:** Version bump only for package @useblu/ocean-react-formik
+
 ## [1.38.2](https://github.com/ocean-ds/ocean-web/compare/v1.38.1...v1.38.2) (2022-12-01)
 
 **Note:** Version bump only for package @useblu/ocean-react-formik